Authenticated with Passport-JWT-Strategy

+++ controllers/api/v1/posts_api.js
@@ -29,16 +29,14 @@ module.exports.index = async function(req , res){
 }
 
 
-
-
-
+// we are authenticating it
 // action for deleting a post
 module.exports.destroy = async function(req , res){
 
     try{
         let post = await Post.findById(req.params.id);
 
-        // if(post.user == req.user.id){
+        if(post.user == req.user.id){
             post.remove();
             await Comment.deleteMany({post : req.params.id});
 
@@ -46,22 +44,12 @@ module.exports.destroy = async function(req , res){
                 message: "Post and assosciated comments deleted!"
             });
 
-            // if(req.xhr){
-            //     return res.status(200).json({
-            //         data:{
-            //             post_id: req.params.id
-            //         },
-            //         message: "Post deleted!"
-            //     });
-            // }
-
-            // req.flash('success' , 'Post and assosciated comments deleted!');
-            // return res.redirect('back');
-        // }else
-        // {
-        //     req.flash('error' , 'you cannot delete this Post');
-        //     return res.redirect('back');
-        // }
+        }else
+        {
+            return res.json(401 , {
+                message: "you can't delete this post!"
+            });
+        }
 
     }catch(err){
         console.log('********' , err);
@@ -71,3 +59,46 @@ module.exports.destroy = async function(req , res){
     }
     
 }
+
+
+
+// // action for deleting a post
+// module.exports.destroy = async function(req , res){
+
+//     try{
+//         let post = await Post.findById(req.params.id);
+
+//         // if(post.user == req.user.id){
+//             post.remove();
+//             await Comment.deleteMany({post : req.params.id});
+
+//             return res.json(200 , {
+//                 message: "Post and assosciated comments deleted!"
+//             });
+
+//             // if(req.xhr){
+//             //     return res.status(200).json({
+//             //         data:{
+//             //             post_id: req.params.id
+//             //         },
+//             //         message: "Post deleted!"
+//             //     });
+//             // }
+
+//             // req.flash('success' , 'Post and assosciated comments deleted!');
+//             // return res.redirect('back');
+//         // }else
+//         // {
+//         //     req.flash('error' , 'you cannot delete this Post');
+//         //     return res.redirect('back');
+//         // }
+
+//     }catch(err){
+//         console.log('********' , err);
+//         // req.flash('error' , err);
+//         return res.json(500 , {
+//             message: "Internal Server Error"
+//         })
+//     }
+    
+// }
